Got email working. Gmail doesn't seem to parse tables (in fact gets rid of table elements), so just using <b>, <u>, &nbsp; Pretty basic HTML.
Reversed order of runs in stats gallery so newest runs are first.

Fixed bug with FunRunOverlay that I introduced when I simplified things.

Fixed GoogleLeg so that it doesn't compute the directions path points every time you call getPathPoints().
Only need to compute this once, when the polyLine is set.

Gah, too tired to go running today for testing. Want to put on market ASAP.